+++ packages/mdx/src/config.ts
@@ -1,65 +1,18 @@
 import path from 'node:path';
 import type { NextConfig } from 'next';
-import {
-  rehypeCode,
-  remarkGfm,
-  remarkStructure,
-  remarkHeading,
-  type RehypeCodeOptions,
-  remarkImage,
-  type RemarkImageOptions,
-  type RemarkHeadingOptions,
-} from 'fumadocs-core/mdx-plugins';
-import type { Pluggable } from 'unified';
 import type { Configuration } from 'webpack';
-import remarkMdxExport from './mdx-plugins/remark-exports';
-import type { LoaderOptions } from './loader';
-import type { MDXLoaderOptionsInput, MDXLoaderOptions } from './loader-mdx';
+import { type InputMDXOptions } from '@/loader-mdx';
 import {
   SearchIndexPlugin,
   type Options as SearchIndexPluginOptions,
-} from './webpack-plugins/search-index-plugin';
-import { RootMapFile } from './root-map-file';
-
-type MDXOptions = Omit<
-  NonNullable<MDXLoaderOptionsInput>,
-  'rehypePlugins' | 'remarkPlugins'
-> & {
-  rehypePlugins?: ResolvePluginsInput;
-  remarkPlugins?: ResolvePluginsInput;
-
-  /**
-   * The folder from which to import the MDX components
-   */
-  providerImportSource?: string | null;
-
-  /**
-   * Properties to export from `vfile.data`
-   */
-  valueToExport?: string[];
-
-  remarkHeadingOptions?: RemarkHeadingOptions;
-  remarkImageOptions?: RemarkImageOptions | false;
-  rehypeCodeOptions?: RehypeCodeOptions | false;
-};
-
-/**
- * A delayed plugin resolution, which dynamically imports the named plugin,
- * and then passes the options to it.
- */
-type LazyPluginResolution = [string, object];
-
-/**
- * A plugin can either be a valid unified plugin or a pair of plugin name and options
- */
-export type ResolvePluginsInput = ResolvePlugins | LazyPluginResolution[];
-
-type ResolvePlugins = Pluggable[] | ((v: Pluggable[]) => Pluggable[]);
+} from '@/webpack-plugins/search-index-plugin';
+import { RootMapFile } from '@/root-map-file';
+import type { LoaderOptions } from './loader';
 
 export interface CreateMDXOptions {
   cwd?: string;
 
-  mdxOptions?: MDXLoaderOptionsInput;
+  mdxOptions?: InputMDXOptions;
 
   buildSearchIndex?:
     | Omit<SearchIndexPluginOptions, 'rootContentDir' | 'rootMapFile'>
@@ -83,103 +36,13 @@ export interface CreateMDXOptions {
    * {@link LoaderOptions.include}
    */
   include?: string | string[];
-}
 
-function pluginOption(
-  def: (v: Pluggable[]) => (Pluggable | false)[],
-  options: ResolvePlugins = [],
-): Pluggable[] {
-  const list = def(Array.isArray(options) ? options : []).filter(
-    Boolean,
-  ) as Pluggable[];
-
-  if (typeof options === 'function') {
-    return options(list);
-  }
-
-  return list;
-}
-
-async function getMDXLoaderOptions({
-  valueToExport = [],
-  rehypeCodeOptions,
-  remarkImageOptions,
-  remarkHeadingOptions,
-  ...mdxOptions
-}: MDXOptions): Promise<MDXLoaderOptions> {
-  const mdxExports = [
-    'structuredData',
-    'toc',
-    'frontmatter',
-    'lastModified',
-    ...valueToExport,
-  ];
-
-  const remarkOptions = await resolveImportPlugins(
-    mdxOptions.remarkPlugins ?? [],
-  );
-  const rehypeOptions = await resolveImportPlugins(
-    mdxOptions.rehypePlugins ?? [],
-  );
-
-  const remarkPlugins = pluginOption(
-    (v) => [
-      remarkGfm,
-      [remarkHeading, remarkHeadingOptions],
-      remarkImageOptions !== false && [remarkImage, remarkImageOptions],
-      ...v,
-      remarkStructure,
-      [remarkMdxExport, { values: mdxExports }],
-    ],
-    remarkOptions,
-  );
-
-  const rehypePlugins = pluginOption(
-    (v) => [
-      rehypeCodeOptions !== false && [rehypeCode, rehypeCodeOptions],
-      ...v,
-    ],
-    rehypeOptions,
-  );
-
-  return {
-    providerImportSource: 'next-mdx-import-source-file',
-    ...mdxOptions,
-    remarkPlugins,
-    rehypePlugins,
-  };
-}
-
-/**
- * When building with turbo we cannot pass closures across the nextjs → turbo boundary,
- * so instead we support the syntax from {@link LazyPluginResolution} which this function
- * then evaluates returning a fully resolved plugin
- */
-async function resolveImportPlugins(
-  input: ResolvePluginsInput,
-): Promise<ResolvePlugins> {
-  return Array.isArray(input)
-    ? await Promise.all(
-        input.map(async (v) => {
-          if (isLazyPluginResolution(v)) {
-            const [pluginName, options] = v;
-            const plugin = (
-              (await import(pluginName)) as {
-                default: (options?: object) => Pluggable;
-              }
-            ).default;
-            return (() => plugin(options)) as Pluggable;
-          }
-          return v;
-        }),
-      )
-    : input;
-}
-
-function isLazyPluginResolution(
-  v: LazyPluginResolution | Pluggable,
-): v is LazyPluginResolution {
-  return Array.isArray(v) && typeof v[0] === 'string';
+  /**
+   * Support Turbopack (experimental)
+   *
+   * @defaultValue false
+   */
+  experimentalTurbo?: boolean;
 }
 
 const defaultPageExtensions = ['mdx', 'md', 'jsx', 'js', 'tsx', 'ts'];
@@ -190,6 +53,7 @@ function createMDX({
   rootMapPath = './.map.ts',
   rootContentPath = './content',
   buildSearchIndex = false,
+  experimentalTurbo = false,
   ...loadOptions
 }: CreateMDXOptions = {}) {
   const rootMapFile = path.resolve(cwd, rootMapPath);
@@ -200,37 +64,48 @@ function createMDX({
       rootMapFile,
     }).create()
   ) {
-    console.log(`Created ${rootMapFile} automatically for you.`);
+    console.log(`Created ${rootMapFile} automatically.`);
   }
 
   return (nextConfig: NextConfig = {}): NextConfig => {
+    if (experimentalTurbo) {
+      if (process.env.NODE_ENV === 'development' && !isSerializable(mdxOptions)) {
+        console.warn('`mdxOptions` must be serializable (e.g. JavaScript primitives)');
+      }
+
+      return {
+        ...nextConfig,
+        pageExtensions: nextConfig.pageExtensions ?? defaultPageExtensions,
+        experimental: {
+          turbo: {
+            rules: {
+              '*.{md,mdx}': [
+                {
+                  loader: 'fumadocs-mdx/loader-mdx',
+                  // TODO: how do we communicate to the user about this?
+                  // @ts-expect-error(arlyon): user must ensure only JSON is sent
+                  options: mdxOptions as object,
+                },
+              ],
+              '.map.ts': [
+                {
+                  loader: 'fumadocs-mdx/loader',
+                  options: {
+                    rootContentDir,
+                    rootMapFile,
+                    ...loadOptions,
+                  } satisfies LoaderOptions,
+                },
+              ],
+            },
+          },
+        },
+      };
+    }
+
     return {
       ...nextConfig,
       pageExtensions: nextConfig.pageExtensions ?? defaultPageExtensions,
-      experimental: {
-        turbo: {
-          rules: {
-            '*.{md,mdx}': [
-              {
-                loader: 'fumadocs-mdx/loader-mdx',
-                // TODO: how do we communicate to the user about this?
-                // @ts-expect-error(arlyon): user must ensure only JSON is sent
-                options: mdxOptions,
-              },
-            ],
-            '.map.ts': [
-              {
-                loader: 'fumadocs-mdx/loader',
-                options: {
-                  rootContentDir,
-                  rootMapFile,
-                  ...loadOptions,
-                },
-              },
-            ],
-          },
-        },
-      },
       webpack: (config: Configuration, options) => {
         config.resolve ||= {};
 
@@ -287,7 +162,21 @@ function createMDX({
   };
 }
 
-export {
-  createMDX as default,
-  getMDXLoaderOptions,
-};
+type JSONValue = string | boolean | number | JSONValue[] | {
+  [key: string]: JSONValue
+}
+
+// TODO: Improve the check function
+function isSerializable(options: unknown): options is JSONValue {
+  if (options === null) return true;
+
+  if (typeof options === 'object') {
+    return Object.values(options).every(isSerializable)
+  }
+
+  if (Array.isArray(options)) options.every(isSerializable)
+
+  return ['string', 'number', 'bigint', 'boolean'].includes(typeof options)
+}
+
+export { createMDX as default };
